Fix and reenable all compiler warnings (#160)
This PR includes a number of backend / build / CI fixes, including:

* Cleanup of many compiler warnings that had previously been swept under the rug
* Targeted suppression of several Boost warnings that can safely be ignored
* Disable `-Werror` for user builds
* Streamlined CI system for easier integration into private repo

+++ core/src/G3MultiFileWriter.cxx
@@ -104,7 +104,7 @@ G3MultiFileWriter::CheckNewFile(G3FramePtr frame)
 		if (!counter)
 			log_fatal("Could not get stream counter");
 
-		if (counter->characters() > size_limit_)
+		if ((size_t)counter->characters() > size_limit_)
 			start_new_ = true;
 
 		if (newfile_callback_.ptr() != Py_None &&
